fixing with level tests... need more caffine
