+++ lib/Ball.js
@@ -1,6 +1,6 @@
-const Render = require('./Render.js');
+const BuildingBlock = require('./BuildingBlock.js');
 
-class Ball extends Render {
+class Ball extends BuildingBlock {
   constructor (x, y, width = 10, height = 10, velocity = 0) {
     super(x, y, width, height);
     this.velocityX = velocity;
