+++ lib/Brick.js
@@ -1,6 +1,6 @@
-const Render = require('./Render.js');
+const BuildingBlock = require('./BuildingBlock.js');
 
-class Brick extends Render {
+class Brick extends BuildingBlock {
   constructor (x, y, width, height, density = 1) {
     super(x, y, width, height);
     this.density = density;
