+++ lib/Level.js
@@ -57,18 +57,21 @@ class Level {
     });
   }
 
+  completeLevel (ball) {
+    ball.velocityX = 0;
+    ball.velocityY = 0;
+    this.currentLevel++;
+    this.bricks = this.buildBricks();
+  }
+
   checkGameState (canvas, context, player, ball) {
     let activeBricks = this.bricks.filter((brick) => {
       return brick.density > 0 && brick.density < 10;
     });
 
     if (activeBricks.length === 0) {
-      ball.velocityX = 0;
-      ball.velocityY = 0;
-      this.currentLevel++;
-      this.bricks = this.buildBricks();
+      this.completeLevel (ball)
     }
-
     if (ball.y > canvas.height) {
       player.removeBall();
       ball.velocityX = 0;
@@ -76,17 +79,21 @@ class Level {
       canvas.classList.add('fail');
     }
     if (player.balls === 0) {
-      this.gameOver = true;
-      context.font = "60px Monospace";
-      context.fillStyle = "#BEC3C1";
-      context.textAlign = "center";
-      context.fillText("Game Over", canvas.width / 2, 300);
-
-      context.font = "20px Monospace";
-      context.fillStyle = "#BEC3C1";
-      context.textAlign = "center";
-      context.fillText("Click to start a new game", canvas.width / 2, 350);
+      this.gameOver(canvas, context);
     }
+  }
+
+  gameOver (canvas, context) {
+    this.gameOver = true;
+    context.font = "60px Monospace";
+    context.fillStyle = "#BEC3C1";
+    context.textAlign = "center";
+    context.fillText("Game Over", canvas.width / 2, 300);
+
+    context.font = "20px Monospace";
+    context.fillStyle = "#BEC3C1";
+    context.textAlign = "center";
+    context.fillText("Click to start a new game", canvas.width / 2, 350);
   }
 
   restartGame (ball, player) {
