+++ lib/Player.js
@@ -1,7 +1,7 @@
-const Render = require('./Render.js');
+const BuildingBlock = require('./BuildingBlock.js');
 const PowerUp = require('./PowerUp.js');
 
-class Player extends Render {
+class Player extends BuildingBlock {
   constructor (x = 450, width = 100, height = 20, score = 0, balls = 3) {
     super(x, 565, width, height);
     this.score = score;
