+++ lib/PowerUp.js
@@ -1,6 +1,6 @@
-const Render = require('./Render.js');
+const BuildingBlock = require('./BuildingBlock.js');
 
-class powerUp extends Render {
+class powerUp extends BuildingBlock {
   constructor (x, y, width, height, player, context, ball) {
     super(x, y, width, height);
     this.powerUpList = [
